Modularize library flag removal and add it to split callback
